% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaPrior.R
\name{alphaPrior}
\alias{alphaPrior}
\title{kummmer}
\usage{
alphaPrior(nMNOfrom, names, variation)
}
\arguments{
\item{nMNOfrom}{}

\item{names}{a list of prior distribution names. It could contain the following names: 'unif', 'degen', 'triang' or 'gamma'}

\item{variation}{}
}
\description{
postestimates of
}
\author{
David Salgado
}
