% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang.R
\name{dtriang}
\alias{dtriang}
\title{The Triangular Distribution.}
\usage{
dtriang(x, xMin, xMax, xMode)
}
\arguments{
\item{x, }{q vector of quantiles}

\item{xMin}{vector with the minimum values of the range of the random variable}

\item{xMax}{vector with the maximum values of the range of the random variable}

\item{xMode}{vector with the modes of the random variable}

\item{p}{vector pf probabilities}

\item{n}{number of observations}
}
\value{
\code{dtriang} gives the density, \code{ptriang} gives the distribution function,
\code{qtriang} gives the quantile function, and \code{rtriang} generates random deviates.

The lengths of the input vectors (except \code{n}) must be all equal except when their length is
1. Otherwise \code{NA}s are produced.
}
\description{
Density, distribution funtion, quantile function and random generation for the
triangular distribution
}
\examples{
curve(dtriang(x, 0, 3, 1), xlim = c(0, 3))
curve(ptriang(x, 0, 3, 1), xlim = c(0, 3))
curve(qtriang(x, 0, 3, 1), xlim = c(0, 1))
hist(rtriang(1e6, 0, 3, 1), breaks = seq(0, 3, by = 0.01))

}
\seealso{
\code{\link{Distributions}} for other distributions
}
