% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genUVLambda.R
\name{genUVLambda}
\alias{genUVLambda}
\title{Generation of three-dimensional random deviates.}
\usage{
genUVLambda(nSim, nStrata, fu, fv, flambda, nMNO)
}
\arguments{
\item{nSim}{number of three-dimensional points to generate}

\item{nStrata}{integer vector of length 3 with the number of strata in each dimension}

\item{fu}{named list with the prior marginal distributions of the parameter \eqn{u}}

\item{fv}{named list with the prior marginal distributions of the parameter \eqn{v}}

\item{flambda}{named list with the prior marginal distributions of the parameter \eqn{\lambda}}

\item{nMNO}{non-negative integer vector with the number of individuals according to the MNO}
}
\value{
\code{genUVLambda} returns a \linkS4class{data.table} with the (u,v, \eqn{\lambda}) 
coordinates of each point together with additional variables:

 \itemize{

   \item The length of \code{nMNO} identifies the number of territorial cells in which the number
    of individuals detected by the telecommunication network. The column \code{cellID} identifies
     these territorial cells.

   \item \code{Stratum_u}, \code{Stratum_v}, and \code{Stratum_lambda} jointly identify each 
   stratum in which the region of integration has been divided with the stratification.
 }
}
\description{
Generate three-dimensional random deviates for a Monte Carlo computation of the
integral \deqn{\int_{0}^{\infty}d\lambda f_{\lambda}(\lambda)\int_{0}^{\infty}dv f_{v}(v)
\int_{0}^{\infty} f_{u}(u)\ \Phi(u\cdot v, (1 - u)
\cdot v; \lambda, N^{\textrm{MNO}}, N^{\textrm{Reg}})\mathbb{P}(N^{\textrm{MNO, rep}}\big|
u,v,\lambda).}

 The Monte Carlo technique makes use of stratified importance sampling.
}
\details{
Notice that \code{nSim} points are generated for each of the \code{length(nMNO)} cells 
so that the final \linkS4class{data.table} has \code{nSim}\eqn{\times}\code{length(nMNO)} rows.

The prior distributions are specified as named lists where the first component of each list must
be the name of distribution ('unif', 'triang', 'gamma') and the rest of components must be
named according to the name of the parameters of the random generator of the corresponding
distribution according to:

  \itemize{

    \item unif: \code{xMin}, \code{xMax} for the minimum, maximum of the sampled interval.
    \item triang: \code{xMin}, \code{xMax}, \code{xMode} for minimum, maximum and mode (see
    \code{\link{qtriang}}).
    \item gamma: \code{scale} and \code{shape} with the same meaning as in \code{\link{rgamma}}.
  }
}
\examples{
# This data.table must have 10x5x3= 150 rows and only one stratum
genUVLambda(nSim = 1000, nStrata = c(1, 2, 5),
      fu = list('unif', xMin = 0.3, xMax = 0.5), 
      fv = list('gamma', shape = 11, scale = 97 / 10),
      flambda = list('gamma', shape = 2, scale = 97 / 1),
      nMNO = c(20, 17, 25))
      
}
\seealso{
\code{\link{genUV}}
}
