% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kummer.R
\name{kummer}
\alias{kummer}
\title{Confluent hypergeometric or Kummer function}
\usage{
kummer(x, a, b, relTol = 1e-06,
  nThreads = RcppParallel::defaultNumThreads())
}
\arguments{
\item{x, }{a, b numeric vectors of the same length}

\item{relTol}{relative tolerance (default value \code{1e-6}) understood as the ratio of each term
in the series relative to the sum}

\item{nThreads}{number (default the number of all cores, including logical cores) to use for computation}
}
\value{
Return a numeric vector with the values of the function
}
\description{
Partial implementation of the confluent hypergeometric function
\eqn{{}_{1}F_{1}(x ; a; b)}
}
\details{
This function is implemented in C++. It is based on Pearson et al (2016). It only
implements the Taylor series method together with an asymtoptic expansion based on Watson's lemma
}
\author{
Luis Sanguiao Bogdan Oancea
}
