% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCheckInd.R
\name{modelCheckInd}
\alias{modelCheckInd}
\title{Indicators for model checking.}
\usage{
modelCheckInd(nSimPar, nMNO, nReg, fu, fv, flambda, relTol = 1e-06,
  nSim = 1e+06, nStrata = c(1, 100, 100), verbose = FALSE,
  nThreads = RcppParallel::defaultNumThreads())
}
\arguments{
\item{nSimPar}{number of simulations to compute the underlying integrals}

\item{nMNO}{non-negative integer vectors with the number of individuals detected according to the
network operator}

\item{nReg}{non-negative integer vectors with the number of individuals detected according to the
population register}

\item{fu}{named list with the prior marginal distribution of the parameter \code{u}}

\item{fv}{named list with the prior marginal distributions of the parameter \code{v}}

\item{flambda}{named list with the prior distribution of the parameter \eqn{\lambda}}

\item{relTol}{relative tolerance in the computation of the \code{\link{kummer}} function. Default
value is \code{1e-6}}

\item{nSim}{number of two-dimensional points to generate to compute the integral. Default value
is \code{1e6}}

\item{nStrata}{integer vector of length 3 with the number of strata in each dimension. Default
value is \code{c(1, 1e2, 1e2)}}

\item{verbose}{logical (default \code{FALSE}) to report progress of the computation}

\item{nThreads}{number (default the number of all cores, including logical cores) to use for computation}
}
\value{
\code{modelCheckInd} returns a \linkS4class{data.table} with the values of the indicators
for model checking:

 \itemize{

   \item nMNO: 

   \item nReg:

   \item B:

   \item relB:
   
   \item V:
   
   \item relV:
   
   \item MSE:
   
   \item relMSE:

 }
}
\description{
Compute the indicators for model checking given by ...
}
\details{
The underlying integrals are computed using with Monte Carlo techniques using 
\code{nSimPar} points for each of the generated random deviate values.

The prior distributions are specified as named lists where the first component of each list must
be the name of distribution ('unif', 'triang', 'degen', 'gamma') and the rest components must be
named according to the name of the parameters of the random generator of the corresponding
distribution according to:

  \itemize{

    \item unif: \code{xMin}, \code{xMax} for the minimum, maximum of the sampled interval.
    \item triang: \code{xMin}, \code{xMax}, \code{xMode} for minimum, maximum and mode (see
    \code{\link{qtriang}}).
    \item gamma: \code{scale} and \code{shape} with the same meaning as in \code{\link{rgamma}}.
  }
}
\examples{
# Easily, a function to draw conditioned on the parameters:
modelCheckInd(nSimPar = 10, nMNO = 29, nReg = 123, 
         fu = list('unif', xMin = 0.2, xMax = 0.25),
         fv = list('unif', xMin = 115, xMax = 130),
         flambda = list('gamma', shape = 21, scale = 123 / 20))

modelCheckInd(nSimPar = 10, nMNO = c(29, 31), nReg = c(123, 119), 
         fu = list(list('unif', xMin = 0.2, xMax = 0.25),
                   list('unif', xMin = 0.21, xMax = 0.26)),
         fv = list(list('unif', xMin = 115, xMax = 130),
                   list('unif', xMin = 114, xMax = 131)),
         flambda = list(list('gamma', shape = 21, scale = 123 / 20),
                        list('gamma', shape = 11, scale = 124 / 10)))
                        

}
\references{
\url{https://github.com/MobilePhoneESSnetBigData}
}
\seealso{
\code{\link{rNMNOrep}}
}
