% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nMNO_ini.R
\docType{data}
\name{nMNO_ini}
\alias{nMNO_ini}
\title{Counts of individuals for the initial time period detected by the mobile network operator.}
\format{A vector with 12 components.}
\usage{
nMNO_ini
}
\description{
This vector contains the counts of individuals for the initial time period detected by the mobile
network operator in each of the 12 cells.
}
\keyword{datasets}
