% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nReg.R
\docType{data}
\name{nReg}
\alias{nReg}
\title{Population counts for the initial time period according to the population register.}
\format{A vector with 12 components.}
\usage{
nReg
}
\description{
This vector contains the population counts for the initial time period according to the
population register.
}
\keyword{datasets}
