% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postN0.R
\name{postN0}
\alias{postN0}
\title{Posterior mean, median, and mode for the number of individuals at the initial time.}
\usage{
postN0(nMNO, nReg, fu, fv, flambda, n = 1000, scale = 1, relTol = 1e-08,
  nSim = 1000, nStrata = c(1, 100), verbose = FALSE,
  nThreads = RcppParallel::defaultNumThreads())
}
\arguments{
\item{nMNO, }{nReg non-negative integer vectors with the number of individuals detected in each
cell according to the network operator and the register}

\item{fu, }{fv named lists with the prior marginal distributions of the two-dimensional points
for the Monte Carlo integration}

\item{flambda}{named list with the prior distribution of the lambda parameter}

\item{n}{number of points to generate in the posterior distribution for the computation. Default
value is 1e3}

\item{scale}{numeric vector with the scale to count the number of individuals. Default value is 1}

\item{relTol}{relative tolerance in the computation of the \code{\link{kummer}} function. Default
value is \code{1e-6}}

\item{nSim}{number of two-dimensional points to generate to compute the integral. Default value
is \code{1e4}}

\item{nStrata}{integer vector of length 2 with the number of strata in each dimension. Default
value is \code{c(1, 1e2)}}

\item{verbose}{logical (default \code{FALSE}) to report progress of the computation}

\item{nThreads}{number (default the number of all cores, including logical cores) to use for computation}
}
\value{
\code{postN0} computes the posterior mean, median, and mode of the posterior distribution
for each cell. The function returns a matrix with the estimates in columns and the cells in rows.
}
\description{
Compute the posterior mean, median, and mode for the number of individuals
generating posterior distribution according to the hierarchical model at the initial time instant
}
\details{
The prior distributions are specified as named lists where the first component of each
list must be the name of distribution ('unif', 'triang', 'degen', 'gamma') and the rest of
components must be named according to the name of the parameters of the random generator of the
corresponding distribution according to:

  \itemize{

    \item unif: \code{xMin}, \code{xMax} for the minimum, maximum of the sampled interval.
    \item degen: \code{x0} for the degenerate value of the random variable.
    \item triang: \code{xMin}, \code{xMax}, \code{xMode} for minimum, maximum and mode (see
    \code{\link{qtriang}}).
    \item gamma: \code{scale} and \code{shape} with the same meaning as in \code{\link{rgamma}}.
  }
}
\examples{
# It takes a couple of minutes
postN0(nMNO = 20, nReg = 115, fu = list('unif', xMin = 0.3, xMax = 0.5),
        fv = list('unif', xMin = 100, xMax = 120),
        flambda = list('gamma', shape = 11, scale = 12))

}
\seealso{
\code{\link{rN0}}
}
