% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postNestimates.R
\name{postNestimates}
\alias{postNestimates}
\title{postNestimates}
\usage{
postNestimates(nMNO, nReg, fu, fv, flambda, n = 1000, relTol = 1e-04,
  nSim = 1000, nStrata = c(1, 100), verbose = FALSE)
}
\author{
David Salgado
}
