% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postNt.R
\name{postNt}
\alias{postNt}
\title{Time evolution of the posterior mean, median, and mode for the number of individuals unconditional on their estimated initial values}
\usage{
postNt(nMNOmat, nReg, fu, fv, flambda, distNames, variation, scale = 1,
  n = 1000, relTol = 1e-05, nSim = 1000, nStrata = c(1, 100),
  verbose = FALSE)
}
\arguments{
\item{nMNOmat}{??}

\item{nReg}{??}

\item{fu, }{fv named lists with the prior marginal distributions of the two-dimensional points
for the Monte Carlo integration}

\item{flambda}{named list with the prior distribution of the lambda parameter}

\item{scale}{numeric vector with the scale to count the number of individuals. Default value is 1}

\item{n}{number of points to generate in the posterior distribution for the computation. Default
value is 1e3}

\item{relTol}{relative tolerance in the computation of the \code{\link{kummer}} function. Default
value is \code{1e-5}}

\item{nSim}{number of two-dimensional points to generate to compute the integral. Default value
is \code{1e3}}

\item{nStrata}{integer vector of length 2 with the number of strata in each dimension. Default
value is \code{c(1, 1e2)}}

\item{verbose}{logical (default \code{FALSE}) to report progress of the computation}
}
\value{
\code{postNt} computes the posterior mean, median, and mode of the posterior distribution
for each cell. The function returns a matrix with the estimates in columns and the cells in rows.
}
\description{
Compute the time evolution of the posterior mean, median, and mode for the number of individuals
unconditional on their estimated initial values, but starting from the input data themselves.
}
\details{
The prior distributions are specified as named lists where the first component of each list must
be the name of distribution ('unif', 'triang', 'degen', 'gamma') and the rest components must be
named according to the name of the parameters of the random generator of the corresponding
distribution according to:

  \itemize{

    \item unif: \code{xMin}, \code{xMax} for the minimum, maximum of the sampled interval.
    \item degen: \code{x0} for the degenerate value of the random variable.
    \item triang: \code{xMin}, \code{xMax}, \code{xMode} for minimum, maximum and mode (see
    \code{\link{qtriang}}).
    \item gamma: \code{scale} and \code{shape} with the same meaning as in \code{\link{rgamma}}.
  }
}
\examples{

}
\seealso{
\code{\link{rNt}}
}
\author{
David Salgado
}
