% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postNtcondN0.R
\name{postNtcondN0}
\alias{postNtcondN0}
\title{Time evolution of the posterior mean, median, and mode for the number of individuals conditional on their estimated initial values}
\usage{
postNtcondN0(N0, nMNOmat, distNames, variation, n = 1000)
}
\arguments{
\item{N0}{initial number of population in each cell}

\item{nMNOmat}{}

\item{distNames}{}

\item{variation}{}

\item{n}{number of points to generate in the posterior distribution for the computation. Default
value is 1e3}
}
\value{

}
\description{
Compute the time evolution of the posterior mean, median, and mode for the number of individuals
conditional on their estimated initial values, according to the hierarchical model.
}
\author{
David Salgado
}
