% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg.R
\name{rg}
\alias{rg}
\title{Generation of random deviates of the candidate distribution.}
\usage{
rg(n, nMNO, nReg, fu, fv, flambda, relTol = 1e-06, nSim = 10000,
  nStrata = c(1, 100), verbose = FALSE)
}
\arguments{
\item{n}{number of values to generate}

\item{nMNO, }{nReg non-negative integer vectors with the number of individuals detected in each
cell according to the network operator and the register}

\item{fu, }{fv named lists with the prior marginal distributions of the two-dimensional points
for the Monte Carlo integration}

\item{flambda}{named list with the prior distribution of the lambda parameter}

\item{relTol}{relative tolerance in the computation of the \code{\link{kummer}} function. Default
value is \code{1e-6}}

\item{nSim}{number of two-dimensional points to generate to compute the integral. Default value
is \code{1e4}}

\item{nStrata}{integer vector of length 2 with the number of strata in each dimension. Default
value is \code{c(1, 1e2)}}

\item{verbose}{logical (default \code{FALSE}) to report progress of the computation}
}
\value{
\code{rg} generates \code{n} points according to the candidate distribution.
}
\description{
Generate random points according to the candidate probability distribution in the
accept-reject method of generation of random variables applied to the distribution of the lambda
parameter
}
\details{
The candidate distribution is a gamma distribution with parameters shape = \code{nMNO} +
1 and scale = \eqn{\lambda^{*}} / \code{nMNO}, where \eqn{\lambda^{*}} stands for the mode of the
posterior distribution of the lambda parameter.

It is important to know that currently this function accepts only parameters for a single cell at
a time. In case of interest for the candidate density function values for a set of cells, the
user should program his/her own routine to apply this function to every cell.
}
\examples{
hist(rg(1e6, nMNO = 20, nReg = 115, fu = list('unif', xMin = 0.3, xMax = 0.5),
        fv = list('unif', xMin = 100, xMax = 120),
        flambda = list('gamma', shape = 11, scale = 12)), breaks = seq(1, 200, by = 2), main ='')

}
\seealso{
\code{\link{modeLambda}}, \code{\link{dlambda}} for related functions.
}

