% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlambda.R
\name{rlambda}
\alias{rlambda}
\title{Generation of random deviates of the posterior distribution of parameter lambda.}
\usage{
rlambda(n, nMNO, nReg, fu, fv, flambda, relTol = 1e-06, nSim = 10000,
  nStrata = c(1, 100), verbose = FALSE,
  nThreads = RcppParallel::defaultNumThreads())
}
\arguments{
\item{n}{number of values to generate}

\item{nMNO, }{nReg non-negative integer vectors with the number of individuals detected in each
cell according to the network operator and the register}

\item{fu, }{fv named lists with the prior marginal distributions of the two-dimensional points
for the Monte Carlo integration}

\item{flambda}{named list with the prior distribution of the lambda parameter}

\item{relTol}{relative tolerance in the computation of the \code{\link{kummer}} function. Default
value is \code{1e-6}}

\item{nSim}{number of two-dimensional points to generate to compute the integral. Default value
is \code{1e4}}

\item{nStrata}{integer vector of length 2 with the number of strata in each dimension. Default
value is \code{c(1, 1e2)}}

\item{verbose}{logical (default \code{FALSE}) to report progress of the computation}
}
\value{
\code{rlambda} generates \code{n} points according to the posterior distribution of
the parameter lambda. The function returns a vector with these points.
}
\description{
Generate random points according to the posterior probability distribution of the
parameter lambda in the hierarchical model.
}
\details{
The points are generated according to the accept-reject method using as candidate
distribution a Cauchy distribution whose parameters are taken from the prior distributions and
the mode of the posterior distribution of the lambda parameter.

The prior distributions are specified as named lists where the first component of each list must
be the name of distribution ('unif', 'triang', 'degen', 'gamma') and the rest components must be
named according to the name of the parameters of the random generator of the corresponding
distribution according to:

  \itemize{

    \item unif: \code{xMin}, \code{xMax} for the minimum, maximum of the sampled interval.
    \item degen: \code{x0} for the degenerate value of the random variable.
    \item triang: \code{xMin}, \code{xMax}, \code{xMode} for minimum, maximum and mode (see
    \code{\link{qtriang}}).
    \item gamma: \code{scale} and \code{shape} with the same meaning as in \code{\link{rgamma}}.
  }
}
\examples{
# It takes a couple of minutes
hist(rN0(500, nMNO = 20, nReg = 115, fu = list('unif', xMin = 0.3, xMax = 0.5),
        fv = list('unif', xMin = 100, xMax = 120),
        flambda = list('gamma', shape = 11, scale = 12))$N0,
        breaks = seq(1, 200, by = 1), main ='', xlab = 'number of individuals')

}
\seealso{
\code{\link{dlambda}}, \code{\link{rg}} for related functions.
}
