% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmatProb.R
\name{rmatProb}
\alias{rmatProb}
\title{rmatProb}
\usage{
rmatProb(n, nMNOmat, distNames, variation)
}
\arguments{
\item{n}{}

\item{nMNOmat}{}

\item{distNames}{a list of prior distribution names. It could contain the following names: 'unif', 'degen', 'triang' or 'gamma'}

\item{variation}{}
}
\value{

}
\description{
postestimates of
}
\author{
David Salgado
}
