% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNMNOrep.R
\name{rNMNOrep}
\alias{rNMNOrep}
\title{Generation random deviates of the replicated number of individuals according to the MNO.}
\usage{
rNMNOrep(n, nMNO, nReg, fu, fv, flambda, relTol = 1e-06, nSim = 1e+06,
  nStrata = c(1, 100, 100), verbose = FALSE,
  nThreads = RcppParallel::defaultNumThreads())
}
\arguments{
\item{n}{number of points to generate}

\item{nMNO}{non-negative integer vectors with the number of individuals detected according to the
network operator}

\item{nReg}{non-negative integer vectors with the number of individuals detected according to the
population register}

\item{fu}{named list with the prior marginal distribution of the parameter \code{u}}

\item{fv}{named list with the prior marginal distributions of the parameter \code{v}}

\item{flambda}{named list with the prior distribution of the parameter \eqn{\lambda}}

\item{relTol}{relative tolerance in the computation of the \code{\link{kummer}} function. Default
value is \code{1e-6}}

\item{nSim}{number of two-dimensional points to generate to compute the integral. Default value
is \code{1e4}}

\item{nStrata}{integer vector of length 3 with the number of strata in each dimension. Default
value is \code{c(1, 1e2, 1e2)}}

\item{verbose}{logical (default \code{FALSE}) to report progress of the computation}

\item{nThreads}{number (default the number of all cores, including logical cores) to use for computation}
}
\value{
\code{rMNOrep} returns a \linkS4class{data.table} with the values of generated points
(column \code{nMNO}) for each value of the parameters \eqn{u, v, \lambda} together with priors 
for the hierarchical model.
}
\description{
Generate random deviates of the number \eqn{N^{\textrm{MNO, rep}}} of individuals 
according to the MNO according to the original number of individuals \eqn{N^{\textrm{MNO}}}
detected by the MNO
}
\details{
The prior distributions are specified as named lists where the first component of each 
list must be the name of distribution ('unif', 'triang', 'gamma') and the rest components must be
named according to the name of the parameters of the random generator of the corresponding
distribution according to:

  \itemize{

    \item unif: \code{xMin}, \code{xMax} for the minimum, maximum of the sampled interval.
    \item degen: \code{x0} for the degenerate value of the random variable.
    \item triang: \code{xMin}, \code{xMax}, \code{xMode} for minimum, maximum and mode (see
    \code{\link{qtriang}}).
    \item gamma: \code{scale} and \code{shape} with the same meaning as in \code{\link{rgamma}}.
  }
}
\examples{

rNMNOrep(10, nMNO = 15, nReg = 110, 
         fu = list('unif', xMin = 0.1, xMax = 0.2),
         fv = list('unif', xMin = 100, xMax = 120),
         flambda = list('gamma', shape = 11, scale = 110 / 10))

}
\references{
\url{https://github.com/MobilePhoneESSnetBigData}
}
\seealso{
\code{\link{rNMNO}}
}
