% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNtcondN0.R
\name{rNtcondN0}
\alias{rNtcondN0}
\title{estimates of the population conditional on their initial values}
\usage{
rNtcondN0(n, N0, nMNOmat, distNames, variation)
}
\description{
estimates of the population conditional on their initial values
}
\author{
David Salgado
}
