% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvlambda.R
\name{ruvlambda}
\alias{ruvlambda}
\title{Generation of random deviates of the posterior distribution of parameters u,v, lambda.}
\usage{
ruvlambda(n, nMNO, nReg, fu, fv, flambda, relTol = 1e-06, nSim = 1e+06,
  nStrata = c(1, 100, 100), verbose = FALSE,
  nThreads = RcppParallel::defaultNumThreads())
}
\arguments{
\item{n}{number of values to generate}

\item{nMNO}{non-negative integer vector with the number of individuals detected in each cell 
according to the network operator}

\item{nReg}{non-negative integer vector with the number of individuals detected in each cell 
according to the register}

\item{fu}{named list with the prior marginal distribution of the parameter \code{u}}

\item{fv}{named list with the prior marginal distributions of the parameter \code{v}}

\item{flambda}{named list with the prior distribution of the parameter \eqn{\lambda}}

\item{relTol}{relative tolerance in the computation of the \code{\link{kummer}} function. Default
value is \code{1e-6}}

\item{nSim}{number of two-dimensional points to generate to compute the integral. Default value
is \code{1e4}}

\item{nStrata}{integer vector of length 3 with the number of strata in each dimension. Default
value is \code{c(1, 1e2, 1e2)}}

\item{verbose}{logical (default \code{FALSE}) to report progress of the computation}

\item{nThreads}{number (default the number of all cores, including logical cores) to use for computation}
}
\value{
\code{ruvlambda} generates \code{n} points according to the posterior distribution of
the parameters \eqn{u, v, \lambda}. The function returns a \linkS4class{data.table} with these 
points.
}
\description{
Generate random points according to the posterior probability distribution of the
parameters u,v, and lambda in the hierarchical model
}
\details{
The points are generated according to the accept-reject method using as candidate
distribution the unnormalised distribution given by 
\deqn{\mathbb{P}(u, v, \lambda|N^{\textrm{MNO}})\propto\mathbb{P}(u,v|\lambda, N^{\textrm{MNO}})
\cdot\mathbb{P}(\lambda|N^{\textrm{MNO}})}.

The prior distributions are specified as named lists where the first component of each list must
be the name of distribution ('unif', 'triang', 'gamma') and the rest components must be
named according to the name of the parameters of the random generator of the corresponding
distribution according to:

  \itemize{

    \item unif: \code{xMin}, \code{xMax} for the minimum, maximum of the sampled interval.
    \item degen: \code{x0} for the degenerate value of the random variable.
    \item triang: \code{xMin}, \code{xMax}, \code{xMode} for minimum, maximum and mode (see
    \code{\link{qtriang}}).
    \item gamma: \code{scale} and \code{shape} with the same meaning as in \code{\link{rgamma}}.
  }
}
\examples{
ruvlambda(10, nMNO = 20, nReg = 115, 
          fu = list('unif', xMin = 0.10, xMax = 0.20),
          fv = list('unif', xMin = 100, xMax = 120),
          flambda = list('gamma', shape = 11, scale = 12))
          
          
ruvlambda(10, nMNO = c(19, 20), nReg = c(115, 117), 
          fu = list(list('unif', xMin = 0.10, xMax = 0.20),
                    list('unif', xMin = 0.11, xMax = 0.22)),
          fv = list(list('unif', xMin = 100, xMax = 120),
                    list('unif', xMin = 97, xMax = 123)),
          flambda = list(list('gamma', shape = 11, scale = 105 / 10),
                         list('gamma', shape = 12, scale = 1107 / 11)))
          
}
\seealso{
\code{\link{rlambda}}, \code{\link{duvlambda}}
}
