% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sim.PopMob}
\alias{Sim.PopMob}
\title{Simulated population mobility.}
\format{A data table with 760 rows and 6 variables:
\describe{
  \item{ID_CELL_INI}{\code{int}, time steps for which we want to estimate the population in each cell}
  \item{ID_CELL_FIN}{\code{int}, time steps for which we want to estimate the population in each cell}
  \item{ID_T}{\code{int}, time steps for which we want to estimate the population in each cell}
  \item{N_REG}{\code{numeric}, the number of population in each cell from the population administrative register}
  \item{N0}{\code{numeric}, the initial number of population in each cell}
  \item{N_MNO_1}{\code{numeric}, the number of population as given by the mobile phone operator}
}}
\usage{
Sim.PopMob
}
\description{
A dataset containing simulated population mobility from one cell to another cell.
}
\keyword{datasets}
