% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang.R
\name{triang}
\alias{triang}
\alias{dtriang}
\alias{ptriang}
\alias{qtriang}
\alias{rtriang}
\title{The Triangular Distribution.}
\usage{
dtriang(x, xMin, xMax, xMode)

ptriang(q, xMin, xMax, xMode)

qtriang(q, xMin, xMax, xMode)

rtriang(n, xMin, xMax, xMode)
}
\arguments{
\item{x, }{q vector of quantiles}

\item{xMin}{the minimum values of the range of the random variable}

\item{xMax}{the maximum values of the range of the random variable}

\item{xMode}{the mode of the random variable}

\item{n}{number of observations}

\item{p}{vector pf probabilities}
}
\value{
\code{dtriang} gives the density, \code{ptriang} gives the distribution function,
\code{qtriang} gives the quantile function, and \code{rtriang} generates random deviates.
}
\description{
Density, distribution funtion, quantile function and random generation for the
triangular distribution
}
\examples{
curve(dtriang(x, 0, 3, 1), xlim = c(0, 3))
curve(ptriang(x, 0, 3, 1), xlim = c(0, 3))
curve(qtriang(x, 0, 3, 1), xlim = c(0, 1))
hist(rtriang(1e6, 0, 3, 1), breaks = seq(0, 3, by = 0.01))

}
\seealso{
\code{\link{Distributions}} for other distributions
}
