% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TruePop}
\alias{TruePop}
\title{True population in each cell at each time step.}
\format{A data table with 60 rows and 3 variables:
\describe{
  \item{Time}{\code{int}, time steps for which we want to estimate the population in each cell}
  \item{Cell}{\code{factor}, the cell number for which we estimate the population}
  \item{N0True}{\code{numeric}, the true population in each cell, at each time moment}
}}
\usage{
TruePop
}
\description{
A dataset containing the true population for a set of cells for succesive moments of time.
}
\keyword{datasets}
