% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNt.R
\name{rNt}
\alias{rNt}
\title{Generation of random deviates of the posterior distribution of population counts.}
\usage{
rNt(n, nMNOmat, nReg, fu, fv, flambda, distNames, variation, scale = 1,
  relTol = 1e-06, nSim = 1000, nStrata = c(1, 100), verbose = FALSE,
  nThreads = RcppParallel::defaultNumThreads())
}
\arguments{
\item{n}{number of values to generate}

\item{nMNOmat}{transition matrix with the number of individuals displaced from cell to cell
detected by the Mobile Network Operator}

\item{nReg}{non-negative integer vectors with the number of individuals detected in each
cell according to the network operator and the register}

\item{fu, }{fv named lists with the prior marginal distributions of the two-dimensional points
for the Monte Carlo integration}

\item{flambda}{named list with the prior distribution of the lambda parameter}

\item{distNames}{character vector with the names of the prior distributions for each cell}

\item{variation}{list of lists whose components are parameters providing a measure of variation
of each prior distribution}

\item{scale}{numeric vector with the scale to count the number of individuals. Default value is 1}

\item{relTol}{relative tolerance in the computation of the \code{\link{kummer}} function. Default
value is \code{1e-6}}

\item{nSim}{number of two-dimensional points to generate to compute the integral. Default value
is \code{1e4}}

\item{nStrata}{integer vector of length 2 with the number of strata in each dimension. Default
value is \code{c(1, 1e2)}}

\item{verbose}{logical (default \code{FALSE}) to report progress of the computation}

\item{nThreads}{number (default the number of all cores, including logical cores) to use for computation}
}
\value{
\code{rNt} generates \code{n} points according to the posterior distribution. The
function returns a \linkS4class{data.table} with these points (under the column \code{N0})
together with the additional variables:

 \itemize{

   \item The common length of \code{nMNO} and \code{nReg} identifies the number of territorial
   cells in which the number of individuals detected by the telecommunication network and
   official data. The column \code{cellID} identifies these territorial cells.

   \item The different values of the generated values of lambda are returned under the column
    \code{lambda}.

   \item The inputs \code{nMNO} and \code{nReg} are also included in the output
   \linkS4class{data.table} in columns under the same name.

 }
}
\description{
Generate random points according to the posterior probability distribution of the
number of individuals in the hierarchical model at arbitrary time instants.
}
\details{
The posterior distribution is a Poisson distribution with parameter \code{lambda *
scale}, where the values of \code{lambda} are generated with the function \code{\link{rlambda}}.

The prior distributions are specified as named lists where the first component of each list must
be the name of distribution ('unif', 'triang', 'degen', 'gamma') and the rest of components must
 be named according to the name of the parameters of the random generator of the corresponding
distribution according to:

  \itemize{

    \item unif: \code{xMin}, \code{xMax} for the minimum, maximum of the sampled interval.
    \item degen: \code{x0} for the degenerate value of the random variable.
    \item triang: \code{xMin}, \code{xMax}, \code{xMode} for minimum, maximum and mode (see
    \code{\link{qtriang}}).
    \item gamma: \code{scale} and \code{shape} with the same meaning as in \code{\link{rgamma}}.
  }
}
\examples{
## First, the inputs:
# The number of generated values
n <- 1e3

#The transition matrix of individuals detected by the MNO
nMNOmat <- rbind(c(10, 3, 4), c(5, 21, 3), c(3, 9, 18))

# Population at the initial time of each cell according to the population register
nReg <- c(90, 130, 101)

# List of priors for u
u0 <- rowSums(nMNOmat) / nReg
cv_u0 <- 0.15
fu <- lapply(u0, function(u){
 umin <- max(0, u - cv_u0 * u)
 umax <- min(1, u + cv_u0 * u)
 output <- list('unif', xMin = umin, xMax = umax)
 return(output)
})

# List of priors for v
v0 <- nReg
cv_v0 <- 0.10
fv <- lapply(v0, function(u){
  umin <- max(0, u - cv_v0 * u)
  umax <- u + cv_v0 * u
  output <- list('unif', xMin = umin, xMax = umax)
  return(output)
})

# List of priors for lambda
cv_lambda <- 0.6
alpha <- 1 / cv_lambda**2 - 1
flambda <- lapply(v0, function(v){list('gamma', shape = 1 + alpha, scale = v / alpha)})

# Names and parameters of priors for the transition probabilities
distNames <- rep('unif', 3)
variation <- rep(list(list(cv = 0.20)), 3)

# The output
Nt <- rNt(n, nMNOmat, nReg, fu, fv, flambda, distNames, variation)$N
hist(Nt, breaks = seq(1, max(Nt) + 10, by = 1), main ='', xlab = 'number of individuals')

}
\seealso{
\code{\link{rlambda}}, \code{\link{rg}}, \code{\link{rNt}} for related functions.
}
