% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNtcondN0.R
\name{rNtcondN0}
\alias{rNtcondN0}
\title{estimates of the population conditional on their initial values}
\usage{
rNtcondN0(n, N0, nMNOmat, distNames, variation)
}
\arguments{
\item{n}{number of values to generate}

\item{N0}{a numeric vector with the initial population estimates for each cell}

\item{nMNOmat}{}

\item{distNames}{a list of prior distribution names. It could contain the following names: 'unif', 'degen', 'triang' or 'gamma'}

\item{variation}{}
}
\value{

}
\description{
estimates of the population conditional on their initial values
}
\author{
David Salgado
}

